+++ CHANGELOG.md
@@ -19,7 +19,7 @@ Introduction of a new module.
 ## mini.comment
 
 - FEATURE: Add `options.start_of_line` option which controls whether to recognize as comment only lines without indent.
-- FEATURE: Add `options.ignore_blank_line` option which controls whether to ignore blank lines when adding comment.
+- FEATURE: Add `options.ignore_blank_line` option which controls whether to ignore blank lines.
 - FEATURE: Add `options.pad_comment_parts` option which controls whether to ensure single space pad for comment leaders.
 
 ## mini.doc